test: split core schedule coverage
